feat: a command is sent when the controller has it, not when it is copied
A CAN message expecting no reply was resolved as soon as the transfer loop
took it out of the ring, which is a memcpy on this side of the link. The
caller believed it was sent while it was still queued, so a drop could not
be reported to anyone: by the time the link went away there was nobody
left waiting.

Every outbound command now takes a sequence number, and the transfer that
carries it reports how far it got. The queue is FIFO end to end - commands
leave the ring in order and are written into a transfer in order - so one
number says what became of any number of them, which keeps this off the
per-command path where the move stream lives. A drop reports everything
queued, covering both what was staged into a transfer that never happened
and what never left the ring.

A message expecting no reply now resolves on that, and fails when the
commands it was among were abandoned.

Also corrects Consts.ProtocolVersion, which disagreed with the transfer
engine's own copy of the same number, and reverts both to 7: nothing has
been released against this protocol.

+++ src/DuetControlServer/Link/Native/NativeMethods.cs
@@ -143,7 +143,7 @@ internal static partial class NativeMethods
     /// <param name="length">Length of <paramref name="message"/> in bytes</param>
     /// <returns>Zero on success, non-zero if the outbound ring is full</returns>
     [LibraryImport(LibraryName)]
-    internal static partial int DuetSbc_QueueMessage(IntPtr handle, uint flags, ReadOnlySpan<byte> message, int length);
+    internal static partial long DuetSbc_QueueMessage(IntPtr handle, uint flags, ReadOnlySpan<byte> message, int length);
 
     /// <summary>
     /// Queue a CAN message for transmission
@@ -158,7 +158,7 @@ internal static partial class NativeMethods
     /// <param name="length">Length of <paramref name="payload"/></param>
     /// <returns>Zero on success, non-zero if the outbound ring is full</returns>
     [LibraryImport(LibraryName)]
-    internal static partial int DuetSbc_QueueCanMessage(IntPtr handle, ushort txToken, ushort msgType,
+    internal static partial long DuetSbc_QueueCanMessage(IntPtr handle, ushort txToken, ushort msgType,
         ushort replyType, byte dstAddress, int isResponse, ReadOnlySpan<byte> payload, int length);
 
     /// <summary>
@@ -169,7 +169,7 @@ internal static partial class NativeMethods
     /// <param name="requestId">Request id to report completion against, or 0 for fire-and-forget</param>
     /// <returns>Zero on success, non-zero if the outbound ring is full</returns>
     [LibraryImport(LibraryName)]
-    internal static partial int DuetSbc_QueueEnableCan(IntPtr handle, int enable, uint requestId);
+    internal static partial long DuetSbc_QueueEnableCan(IntPtr handle, int enable, uint requestId);
 
     /// <summary>
     /// Request an immediate emergency stop
